mejoras visuales solicitadas
